pping: Add option to configure XDP mode
Add an option to let the user configure which mode to load the XDP
program in (unspecified, native or generic).

Set the default mode to native (was unspecified previously) as that is
what the user most likely wants to use (generic or unpsecified falling
back on generic will likely have worse performance).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>