Take off the XDP_ZEROCOPY flag from port_params_default.xsk_cfg.bind_flags
Fix for issue 78; veth does not support zerocopy in bind flags
Take off the XDP_ZEROCOPY flag in the setting of
port_params_default.xsk_cfg.bind_flags in AF_XDP-forwarding/xsk_fwd.c

With this change, libxdp first tries to set up zerocopy, and when it finds
that this is not available it sets up an implementation which copies the
data. So performance will not be impactes for eths which support zerocopy.

Signed-off-by: Chris Ward <tjcw@uk.ibm.com>