pping: Add script for cleaing up leftover tc programs
In case ePPing is not shut down cleanly (ex. when killed with SIGKILL
or OOM killer) it will not be able to detach its eBPF programs, may
remain attached and waste resources. Add a script which can be used
to clean up any remaining programs tc-eBPF programs.

Note, this script should not be run while the any instace of pping is
still running, as that will remove its tc programs and thus its
ability to function properly.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>