pping: Add additional per-prefix packet counters
When running in aggregated mode (-a/--aggregate), the previous
per-prefix packet and byte counters only included traffic that the RTT
was tracked for, i.e. by default only TCP traffic with TCP
timestamps (which a flowstate could be created for) was counted. This
makes it hard to correlate RTTs with traffic load, as the total
traffic load to/from a given prefix is not known.

Therefore, split up the per-prefix counters into 3 sets of counters:
- One for TCP traffic with timestamps, i.e. the TCP traffic we can
  track RTTs of
- One of TCP traffic without timestamps, i.e TCP traffic we cannot
  track due to relying on TCP timestamps
- One for non-TCP traffic, which when combined with the other counters
  gives the total amount of (IP) traffic going to/from a prefix

Do NOT create NEW prefix entries for traffic which the RTT cannot be
tracked for. This means that if some prefix only sees traffic of a
type that RTTs cannot be captured for, they will use the global /0
backup entries.

To keep the standard output somewhat manageable (it is already quite
wide), only output the total packet and byte counts for the traffic
to/from the prefix. For the JSON format, output the counters for each
individual set (TCP_TS, TCP_noTS, and other) which are non-empty.

Example standard entry after update (same as before update, linebreaks
not part of actual output):

14:42:10.451929078: 10.11.1.10/32 -> rxpkts=4303, rxbytes=347742,
txpkts=37658, txbytes=1888184076, rtt-count=1202, min=0.006963 ms,
mean=2 ms, median=2 ms, p95=2 ms, max=3.10063 ms

Example JSON entry before update:

{
  "timestamp": 1697638197074346000,
  "ip_prefix": "10.11.1.10/32",
  "rx_packets": 2495,
  "tx_packets": 12121,
  "rx_bytes": 164670,
  "tx_bytes": 601306338,
  "count_rtt": 743,
  "min_rtt": 7717,
  "mean_rtt": 2021530,
  "median_rtt": 2000000,
  "p95_rtt": 2000000,
  "max_rtt": 4985117,
  "histogram": [
    739,
    4
  ]
}

Example JSON entry after update:

{
  "timestamp": 1697635990442789000,
  "ip_prefix": "10.11.1.10/32",
  "rx_stats": {
    "TCP_TS": {
      "packets": 1458,
      "bytes": 96232
    },
    "TCP_noTS": {
      "packets": 1,
      "bytes": 74
    },
    "other": {
      "packets": 1874,
      "bytes": 183460
    }
  },
  "tx_stats": {
    "TCP_TS": {
      "packets": 17270,
      "bytes": 905414662
    },
    "TCP_noTS": {
      "packets": 1,
      "bytes": 74
    },
    "other": {
      "packets": 1898,
      "bytes": 184204
    }
  },
  "count_rtt": 629,
  "min_rtt": 7775,
  "mean_rtt": 2038160,
  "median_rtt": 2000000,
  "p95_rtt": 2000000,
  "max_rtt": 13431771,
  "histogram": [
    627,
    0,
    0,
    2
  ]
}

This commit will considerably increase the overhead for traffic types
that RTT isn't tracked for compared to before. Previously, the eBPF
programs would early abort as soon as it discovered that the packet
was of a type which it couldn't track RTT for. Now, all IP packets
will have their IP-address processed and later used to lookup the
relevant prefixes to update the packet counters for. However, the
overhead for packets that the RTT can't be tracked for should still be
considerbly lower than for packets the RTT can be tracked for, and the
overhead for packets the RTT can be tracked for should not increase
much from previously.

Potential bug/issue. If the program is configured to NOT track RTTs
for TCP traffic (by using the -I/--icmp flag without the -T/--tcp
flag), the program will not parse the TCP header and thus be unable to
detect if it contains TCP timestamps. Therefore, all TCP packets will
then be counted as TCP packets without timestamps, regardless if they
have timestamps or not.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>