Refactor proto_to_str() function to be thread-safe
Refactor the proto_to_str() function to write the protocol string to a
provided buffer instead of providing a pointer to a static
buffer. This makes it possible to safely use the function in
multi-threaded contexts or call it multiple times before printing the
returned value.

Also rename it to ipproto_to_str(), to clarify that it's for IP
protocols.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>