pping: Add ECN counters to the global counters
Add counters for the 4 ECN code points (00=Not-ECT, 01=ECT1, 10=ECT0
and 11=CE) to the global counters. These are reported together with
the global protocol counters when running in aggregated mode.

Example standard entry (linebreaks not part of actual output):

19:32:40.224309565: non-IP=(pkts=6, bytes=252), UDP=(pkts=9,
bytes=495), ECN=(Not-ECT=4, ECT1=3, CE=2)

Example JSON entry:
{
  "timestamp": 1698082435757528300,
  "protocol_counters": {
    "non_IP": {
      "packets": 6,
      "bytes": 252
    },
    "UDP": {
      "packets": 9,
      "bytes": 495
    }
  },
  "ecn_counters": {
    "no_ECT": 4,
    "ECT1": 3,
    "CE": 2
  }
}

Originally planned to also include a counter for ECN-echo in the TCP
header. However, adding parsing of a TCP field for ALL TCP packets is
currently challenging due to the parsing of the TCP-header being
conditional. First off, the TCP-header will only be parsed if the
program is configured to capture TCP RTTs (can be disabled by passing
the -C/--icmp flag without the -T/--tcp flag). Second off, parsing the
TCP-header is tied to parsing TCP timestamps, and the function for
parsing the TCP timestamps will signal failure regardless if it failed
to parse the TCP-header itself or just the TCP timestamps. Parsing of
a TCP field (like ECE) is thus only safe for the subset of packets
where TCP timestamps could successfully be parsed, which would create
misleading stats as the other ECN counters cover all IP-traffic.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>