netstacklat: Ignore skbs where skb->tstamp is TX timestamp
Since kernel v5.18 the skb->tstamp value for ingress packets may
contain a preserved TX timestamp to later be used on the egress path,
e.g. for EDT, see [1]. Using these TX timestamps to calculate latency
on the ingress path results in erroneous values being reported, and
due to these TX timestamps often using CLOCK_MONOTONIC instead of
CLOCK_REALTIME the error is often years of latency.

To prevent these erroneous values, check if the skb->tstamp is a TX or
RX timetamp, and ignore skbs with TX timestamps.

How to check if skb->tstamp is a TX timestamp depends on kernel
version. There are 3 cases to consider:
- Kernel < v5.18: The skb->tstamp is always reset on ingress, will
  never be a TX timestamp.
- v5.18 <= kernel < v6.11: If the skb->mono_delivery_time field is
  set, the skb->tstamp is a preserved TX timestamp.
- Kernel >= v6.11: The skb->mono_deliver_field has been renamed to
  skb->tstamp_type [2] to support additional clock_ids. If set to
  a non-zero value the skb->tstamp is a TX timestamp.

As we rely on the bpf_ktime_get_tai_ns() helper, which was introduced
in v6.1, we can ignore the first case as kernel < v5.18 will be
unsupported anyways. To handle the renaming of the field in v6.11
utilize CO-RE functionality to determine if the skb->tstamp_type field
exists. Use the bpf_core_field_exists() together with "overloading"
the sk_buff definition using the triple-dash ignored suffix rule as
explained in [3].

[1] https://lore.kernel.org/all/20220302195525.3480280-1-kafai@fb.com/
[2] https://lore.kernel.org/all/20240509211834.3235191-2-quic_abchauha@quicinc.com/
[3] https://nakryiko.com/posts/bpf-core-reference-guide/#handling-incompatible-field-and-type-changes

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>