netstacklat: Add pid filter
Add the -p/--pid-filter option, which allows the user to provide a
comma-separated list of PIDs for applications that the output should
be filtered for.

Only filter hooks running in the process context (currently only the
socket-read hooks), as the receiving application is generally not
known in the earlier parts of the stack. To also filter the earlier
hooks would require keeping track of all connections, so that the
skbs (based on the 5-tuple) can be matched to the receiving
application, and both maintaining this state and looking it up would
add considerable overhead.

To minimize overhead, implement the pid matching as a simple array map
lookup. As the maximum possible PID value is 4M, a full array map
containing every possible PID value takes up around 32MB.

Note that currently no more than 4096 PIDs may be filtered for. This
is only a limit of how many PIDs the user space agent can store when
parsing the PID list, and can easily be extended, as the BPF side can
handle up to 4M PIDs.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>