netstacklat: Generalize initialization of array filter maps
The filtering for specific pids (--pids/-p) makes use of a BPF array
map where each entry to be included is set to 1 (the rest remain 0 as
all entries by default are zeroed in array maps). Generalize the user
space logic that initializes the entries in this filter map so that it
can be reused by other similar features in the future.

Furthermore, change the value type of the filter map from u8 to
u64. While only one bit is really required, each BPF map entry takes
up at least 8 bytes regardless, so using u8 does not save any
space. Additionally, a future commit will add functionality to also
filter by cgroup, and to keep that feature compatible with
ebpf-exporter the value size needs to be 8 bytes. So use u64 keys for
all filter maps from here on out to make it easier to reuse logic.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>