netstacklat: Multiplex all histograms in a single map
Change the way that the latency histograms are stored in BPF
maps. Instead of keeping a separate array map for each histogram,
store all histograms in a single hash map, encoding the hook as part
of the key.

This results in higher overhead (as hash lookups are slower than array
lookups), but is much more flexible. This makes it easier to add
additional hook points as no new maps (and related code for mapping
hooks to maps) need to be added. Furthermore, in the future it allows
to easily group the results on various aspects by adding additional
members to the key.

On the userspace side, maintain a sorted array of the encountered
histogram keys and a mapping to the corresponding histogram
buckets. Instead of keeping a separate key for each histogram
bucket (as the BPF maps do to be compatible with ebpf-exporter),
restructure the data so only a single key is used per
histogram. Essentially remove the bucket member from the key, keeping
a full histogram (where any missing buckets are zeroed) for the
remaining unique members in the histogram key (so far just the hook
identifier).

Keeping the array of histogram keys sorted allows for relatively quick
lookups using binary search. When a new histogram key is encountered
it will incur significant overhead the first time as it needs to be
inserted into the right place in the array, but lookups ought to be
much more common than inserting new keys. While this data structure
will not scale well to a very large amount of unique keys (insertion
time is O(n), lookup O(log n)), it avoids implementing or adding
dependencies to more complicated data structures like trees or
hash maps. As long as we do not need to keep track of many thousands
of histograms, this solution should be good enough.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>