netstacklat: Update socket enqueue hooks
Previously netstacklat used fexit:tcp_data_queue as the socket enqueue
point for TCP and fexit:udp[v6]_queue_rcv_one_skb for UDP. For both of
these functions, the skb may actually have been freed by the time they
return, leading us to read invalid data. Furthermore, not all calls to
these functions necessarily end up enqueuing the skb to the socket,
as they may be dropped for various reasons. For TCP, there are also
some fast paths that may enqueue data to the socket without going
through tcp_data_queue. Therefore, update these probes to hook more
suitable functions.

For TCP, use the tcp_queue_rcv function (which is called by
tcp_data_queue when the data is actually queued to the socket). This
function is much closer to the actual socket enqueue point, will never
free the skb itself (although its return value may indicate to the
calling function that it should be freed as it's be coalesced into
tail skb in the receive queue), is only called when the skb is
actually queued to the socket, and is also called in a fast path of
tcp_rcv_established that bypasses tcp_data_queue.

For UDP, use __udp_enqueue_schedule_skb, which
udp[v6]_queue_rcv_one_skb functions call when they actually attempt to
enqueue the skb to the socket. This function may still fail to enqueue
to skb to the socket (if e.g. the socket buffer is full), so check the
return value so that we only report the instances where the skb is
successfully enqueued. This function is called by both the IPv4 and
IPv6 UDP paths, so similar to the TCP case we only need to hook a
single function now.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>