trie-free-latency: Demo showing the latency of freeing LPM_TRIE maps
The performance of freeing LPM_TRIE maps suffers heavily for maps with a
large number of entries. Here's a little demo that measures the duration
of trie_free() and allows the user to control how many entries are in
the map, as well as how many iterations of the demo are run to collect
latency statistics.

Signed-off-by: Matt Fleming <mfleming@cloudflare.com>