lib/common.mk: Use clang directly to build BPF objects
It turns out we pass both -S and -c to clang, which has become an error
starting from clang 20. The two-stage build using 'clang -S' followed by
llc was used in the early days of BPF development, but these days, clang
is perfectly happy to produce BPF object files directly.

So use this opportunity to get rid of llc entirely, and simplify the BPF
object build rules while we're add it, by moving all of the -W* defines
to BPF_CFLAGS.

This is ported from an identical commit to xdp-tools.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>