configure: Don't fail entirely if the clang errno check fails
Some of the examples compile fine without being able to include errno.h
from BPF code. To allow these to compile, let's make the configure check
a regular feature check, and simply exclude the directories that require
errno.h from the top-level Makefile if the header is not available.

Move the explanation for how to fix the error into README.md, and refer
to that from the configure script output.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>